# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"Returns a welcome message."}},"summary":"Get a welcome message","tags":["App"]}},"/smolagent/interact":{"post":{"operationId":"SmolAgentController_interact","parameters":[],"responses":{"201":{"description":""}},"tags":["SmolAgent"]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"The user has been successfully created."},"400":{"description":"Invalid input."}},"security":[{"bearer":[]}],"summary":"Create a new user","tags":["Users"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"List of all users."}},"security":[{"bearer":[]}],"summary":"Retrieve all users","tags":["Users"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the user to retrieve","schema":{"type":"string"}}],"responses":{"200":{"description":"The user details."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Retrieve a user by ID","tags":["Users"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the user to update","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"The user has been successfully updated."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Update a user by ID","tags":["Users"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the user to delete","schema":{"type":"string"}}],"responses":{"200":{"description":"The user has been successfully deleted."},"403":{"description":"Forbidden. Only admins can access this route."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Delete a user by ID (Admin only)","tags":["Users"]}},"/profile":{"get":{"operationId":"ProfileController_getProfile","parameters":[],"responses":{"200":{"description":"The user profile."},"401":{"description":"Unauthorized."}},"security":[{"bearer":[]}],"summary":"Get the profile of the logged-in user","tags":["Profile"]},"patch":{"operationId":"ProfileController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"The user profile has been successfully updated."},"400":{"description":"Invalid input."},"401":{"description":"Unauthorized."}},"security":[{"bearer":[]}],"summary":"Update the profile of the logged-in user","tags":["Profile"]}},"/profile/username":{"get":{"operationId":"ProfileController_getUsername","parameters":[],"responses":{"200":{"description":"The username of the user."},"401":{"description":"Unauthorized."}},"security":[{"bearer":[]}],"summary":"Get the username of the logged-in user","tags":["Profile"]}},"/profile/admin/users":{"get":{"operationId":"ProfileController_getAllUsers","parameters":[],"responses":{"200":{"description":"List of all users."},"403":{"description":"Forbidden. Only admins can access this route."}},"security":[{"bearer":[]}],"summary":"Get all users (Admin only)","tags":["Profile"]}},"/profile/admin/users/{id}":{"get":{"operationId":"ProfileController_getUserById","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the user to retrieve","schema":{"type":"number"}}],"responses":{"200":{"description":"The user details."},"403":{"description":"Forbidden. Only admins can access this route."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Get a specific user by ID (Admin only)","tags":["Profile"]}},"/cv":{"post":{"operationId":"CvControllerV2_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCvDto"}}}},"responses":{"201":{"description":"The CV has been successfully created."},"400":{"description":"Invalid input."}},"security":[{"bearer":[]}],"summary":"Create a new CV","tags":["CV"]},"get":{"operationId":"CvControllerV2_getCvs","parameters":[{"name":"search","required":false,"in":"query","description":"Search term for filtering CVs","schema":{"example":"Software Engineer","type":"string"}},{"name":"age","required":false,"in":"query","description":"Filter CVs by age","schema":{"example":30,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination (default is 1)","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page (default is 10)","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":"List of CVs."}},"security":[{"bearer":[]}],"summary":"Retrieve CVs with optional filters and pagination","tags":["CV"]}},"/cv/upload/{id}":{"post":{"operationId":"CvControllerV2_uploadCvImage","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the CV to upload the image for","schema":{"type":"number"}}],"responses":{"200":{"description":"The CV image has been successfully uploaded."},"400":{"description":"File is required."}},"security":[{"bearer":[]}],"summary":"Upload an image for a CV","tags":["CV"]}},"/cv/{id}":{"get":{"operationId":"CvControllerV2_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the CV to retrieve","schema":{"type":"string"}}],"responses":{"200":{"description":"The CV details."},"404":{"description":"CV not found."}},"security":[{"bearer":[]}],"summary":"Retrieve a CV by ID","tags":["CV"]},"patch":{"operationId":"CvControllerV2_update","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the CV to update","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCvDto"}}}},"responses":{"200":{"description":"The CV has been successfully updated."},"403":{"description":"Forbidden. You are not allowed to update this CV."},"404":{"description":"CV not found."}},"security":[{"bearer":[]}],"summary":"Update a CV by ID","tags":["CV"]},"delete":{"operationId":"CvControllerV2_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the CV to delete","schema":{"type":"string"}}],"responses":{"200":{"description":"The CV has been successfully deleted."},"403":{"description":"Forbidden. You are not allowed to delete this CV."},"404":{"description":"CV not found."}},"security":[{"bearer":[]}],"summary":"Delete a CV by ID","tags":["CV"]}},"/cv/image/{id}":{"get":{"operationId":"CvControllerV2_getCvImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["CV"]}},"/skill":{"post":{"operationId":"SkillController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSkillDto"}}}},"responses":{"201":{"description":"The skill has been successfully created."},"400":{"description":"Invalid input."}},"security":[{"bearer":[]}],"summary":"Create a new skill","tags":["Skills"]},"get":{"operationId":"SkillController_findAll","parameters":[],"responses":{"200":{"description":"List of all skills."}},"security":[{"bearer":[]}],"summary":"Retrieve all skills","tags":["Skills"]}},"/skill/{id}":{"get":{"operationId":"SkillController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the skill to retrieve","schema":{"type":"string"}}],"responses":{"200":{"description":"The skill details."},"404":{"description":"Skill not found."}},"security":[{"bearer":[]}],"summary":"Retrieve a skill by ID","tags":["Skills"]},"patch":{"operationId":"SkillController_update","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the skill to update","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSkillDto"}}}},"responses":{"200":{"description":"The skill has been successfully updated."},"404":{"description":"Skill not found."}},"security":[{"bearer":[]}],"summary":"Update a skill by ID","tags":["Skills"]},"delete":{"operationId":"SkillController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID of the skill to delete","schema":{"type":"string"}}],"responses":{"200":{"description":"The skill has been successfully deleted."},"404":{"description":"Skill not found."}},"security":[{"bearer":[]}],"summary":"Delete a skill by ID","tags":["Skills"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully."},"400":{"description":"Invalid input."}},"summary":"Register a new user","tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User logged in successfully and access token returned."},"401":{"description":"Invalid credentials."}},"summary":"Login a user","tags":["Auth"]}}},"info":{"title":"CV Manager API","description":"API documentation for the CV Manager project","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"john_doe"},"email":{"type":"string","description":"Email address of the user","example":"john@example.com"},"password":{"type":"string","description":"Password of the user","example":"securePassword123","writeOnly":true}},"required":["username","email","password"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Optional username of the user","example":"john_doe_updated"},"email":{"type":"string","description":"Optional email address of the user","example":"john_updated@example.com"},"password":{"type":"string","description":"Optional password of the user","example":"newSecurePassword123","writeOnly":true}}},"CreateCvDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the CV owner","example":"John"},"firstname":{"type":"string","description":"First name of the CV owner","example":"Doe"},"age":{"type":"number","description":"Age of the CV owner","example":30},"cin":{"type":"string","description":"CIN (National ID) of the CV owner","example":"12345678"},"job":{"type":"string","description":"Job title of the CV owner","example":"Software Engineer"},"path":{"type":"string","description":"Path or URL of the CV image","example":"/uploads/cv-image.jpg"}},"required":["name","firstname","age","cin","job","path"]},"UpdateCvDto":{"type":"object","properties":{"name":{"type":"string","description":"Optional name of the CV owner","example":"John"},"firstname":{"type":"string","description":"Optional first name of the CV owner","example":"Doe"},"age":{"type":"number","description":"Optional age of the CV owner","example":30},"cin":{"type":"string","description":"Optional CIN (National ID) of the CV owner","example":"12345678"},"job":{"type":"string","description":"Optional job title of the CV owner","example":"Software Engineer"},"path":{"type":"string","description":"Optional path or URL of the CV image","example":"/uploads/cv-image.jpg"}}},"CreateSkillDto":{"type":"object","properties":{"designation":{"type":"string","description":"Designation or name of the skill","example":"JavaScript"}},"required":["designation"]},"UpdateSkillDto":{"type":"object","properties":{"designation":{"type":"string","description":"Optional designation or name of the skill","example":"TypeScript"}}},"RegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"john_doe"},"password":{"type":"string","description":"Password of the user","example":"securePassword123"},"email":{"type":"string","description":"Email address of the user","example":"john@example.com"}},"required":["username","password","email"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"john_doe"},"password":{"type":"string","description":"Password of the user","example":"securePassword123"}},"required":["username","password"]}}}}